# openapi.yaml
openapi: 3.0.3
info:
  title: Eskisini Ver Yenisini Al API
  description: |
    Trade-in platformu API'si. Kullanıcıların eski ürünlerini değerlendirip
    yeni ürünlerle takas etmelerine olanak sağlar.

    ## Domain Modeli

    ### CatalogProduct vs TradeInItem
    - **CatalogProduct**: Satışa sunulan yeni/yenilenmiş ürünler (stok, fiyat, garanti)
    - **TradeInItem**: Kullanıcının trade-in için gönderdiği eski cihaz (değerleme, durum, IMEI)

    ### Naming Conventions

    | Domain | Entity | Endpoint | Tag |
    |--------|--------|----------|-----|
    | Ürün Kataloğu | `CatalogProduct` | `/catalog/products` | Catalog |
    | Trade-In | `TradeIn` | `/trade-ins` | TradeIn |
    | Değerleme | `Valuation` | `/trade-ins/valuations` | TradeIn |
    | Sipariş | `Order` | `/orders` | Orders |
    | Kredi | `Credit` | `/credits` | Users |

    ### Terminoloji Standardı
    - **Credits** (not Wallet): Kullanıcı bakiyesi için "Credits" terminolojisi kullanılır
    - **TradeIn** (not Submission): Ana domain ismi "TradeIn", detay entity "TradeIn" olarak kalır
    - **Error**: Tüm hata response'ları `ErrorResponse` şemasını kullanır

    ### State Machine Standardı
    Tüm status alanları SCREAMING_CASE kullanır ve tutarlı geçişlere sahiptir.

    ### Settlement Policy (MVP)
    - **CREDIT**: Trade-in değeri kullanıcı hesabına kredi olarak eklenir
    - **ORDER_DISCOUNT**: Trade-in değeri siparişten düşülür
...

  version: 1.1.0
  contact:
    name: API Destek
...
  license:
...
servers:
...
paths:
...
  /trade-ins/brands:
    get:
      tags:
        - TradeIn
      summary: Trade-in desteklenen markaları listele
      operationId: listTradeInBrands
      security: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Marka listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brand'
        '400':
          $ref: '#/components/responses/BadRequest'
  /trade-ins/models:
    get:
      tags:
        - TradeIn
      summary: Trade-in desteklenen modelleri listele
      operationId: listTradeInModels
      security: []
      parameters:
        - name: brandId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Model listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceModel'
        '400':
          $ref: '#/components/responses/BadRequest'
  /trade-ins/valuations:
    get:
      tags:
        - TradeIn
      summary: Kullanıcının değerleme geçmişini listele
      operationId: listValuations
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ValuationStatus'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Değerleme listesi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TradeIn
      summary: Yeni değerleme talebi oluştur
      description: |
        Kullanıcının eski cihazı için anlık değerleme teklifi al.
        Bu işlem bir TradeInItem oluşturur ve değerleme yapar.
      operationId: createValuation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateValuationRequest'
      responses:
        '201':
          description: Değerleme oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Valuation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /trade-ins/valuations/{valuationId}:
    get:
      tags:
        - TradeIn
      summary: Değerleme detayını getir
      operationId: getValuation
      security:
        - bearerAuth: []
      parameters:
        - name: valuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Değerleme detayı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Valuation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /trade-ins/valuations/{valuationId}/accept:
    post:
      tags:
        - TradeIn
      summary: Değerleme teklifini kabul et ve gönderim başlat
      description: |
        Teklifi kabul eder ve TradeIn oluşturur.
        Status: PENDING → QUOTED → ACCEPTED geçişi yapar.
      operationId: acceptValuation
      security:
        - bearerAuth: []
      parameters:
        - name: valuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptValuationRequest'
      responses:
        '200':
          description: Teklif kabul edildi, gönderim oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeIn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Değerleme zaten kullanılmış veya süresi dolmuş
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trade-ins:
    get:
      tags:
        - TradeIn
      summary: Kullanıcının trade-in gönderimlerini listele
      operationId: listTradeIns
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TradeInStatus'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Trade-in gönderim listesi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeInListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /trade-ins/{tradeInId}:
    get:
      tags:
        - TradeIn
      summary: Trade-in gönderim detayı
      operationId: getTradeIn
      security:
        - bearerAuth: []
      parameters:
        - name: tradeInId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Trade-in detayı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeIn'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Geçersiz istek
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Kimlik doğrulama gerekli
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Kaynak bulunamadı
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ValuationStatus:
      type: string
      enum:
        - PENDING      # Değerleme hesaplanıyor
        - QUOTED       # Teklif hazır, kullanıcı yanıtı bekleniyor
        - ACCEPTED     # Kullanıcı teklifi kabul etti
        - EXPIRED      # Teklif süresi doldu
        - USED         # Değerleme bir siparişte kullanıldı
        - REJECTED     # Kullanıcı teklifi reddetti

    # Trade-In Status: Gönderim durumu
    # CREATED → LABEL_CREATED → IN_TRANSIT → RECEIVED → INSPECTING
    #        → FINAL_OFFERED → COMPLETED/RETURNED
    #        → REJECTED (herhangi bir aşamada)
    TradeInStatus:
      type: string
      enum:
        - CREATED        # Gönderim oluşturuldu
        - LABEL_CREATED  # Kargo etiketi oluşturuldu
        - IN_TRANSIT     # Kargoda, yolda
        - RECEIVED       # Teslim alındı, inceleme bekliyor
        - INSPECTING     # İnceleme yapılıyor
        - FINAL_OFFERED  # Final teklif yapıldı, kullanıcı yanıtı bekleniyor
        - COMPLETED      # İşlem tamamlandı, kredi yüklendi
        - REJECTED       # Ürün kabul edilmedi (BTK sorun, sahte vb.)
        - RETURNED       # Kullanıcı final teklifi reddetti, ürün iade ediliyor

    # Order Status: Sipariş durumu
...
    SettlementType:
      type: string
      x-mvp-only: true
      x-future-values:
        - BANK_TRANSFER  # Faz 2 - KYC/AML sonrası
      enum:
        - CREDIT          # Platform kredisi olarak ekle
        - ORDER_DISCOUNT  # Siparişten düş
      description: |
        **MVP Settlement Seçenekleri (Sadece bu değerler kabul edilir):**
        - `CREDIT`: Trade-in değeri kullanıcı hesabına kredi olarak eklenir
        - `ORDER_DISCOUNT`: Trade-in değeri mevcut/yeni siparişten düşülür

        ⚠️ **BANK_TRANSFER MVP'de YOKTUR.**
        Faz 2'de KYC/AML kontrolleri sonrası eklenecektir.
        Server-side validation bu değeri reddeder.

    # Catalog Product Condition
...
    TradeInItemCondition:
      type: string
      enum:
        - EXCELLENT      # Mükemmel durumda
        - GOOD           # İyi durumda
        - FAIR           # Orta durumda
        - POOR           # Kötü durumda
      description: Kullanıcının trade-in için gönderdiği ürünün durumu

    # ============================================
    # AUTH SCHEMAS
    # ============================================
...
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        logo:
          type: string
          format: uri
        categoryIds:
          type: array
          items:
            type: string
            format: uuid
    DeviceModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
        name:
          type: string
        releaseYear:
          type: integer
        variants:
          type: array
          items:
            $ref: '#/components/schemas/DeviceVariant'
    DeviceVariant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        storage:
          type: string
          example: "128GB"
        color:
          type: string
        baseTradeInValue:
          type: number
          description: EXCELLENT durumda tahmini trade-in değeri (TL)

    # ============================================
    # CATALOG PRODUCT SCHEMAS (Satıştaki Ürünler)
    # ============================================
...
    TradeInItem:
      type: object
      description: Kullanıcının trade-in için gönderdiği eski cihaz
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
        brandName:
          type: string
        modelId:
          type: string
          format: uuid
        modelName:
          type: string
        variantId:
          type: string
          format: uuid
        variantDetails:
          type: string
          example: "128GB, Siyah"
        declaredCondition:
          $ref: '#/components/schemas/TradeInItemCondition'
        actualCondition:
          $ref: '#/components/schemas/TradeInItemCondition'
          description: Fiziksel inceleme sonrası belirlenen durum
        conditionDetails:
          $ref: '#/components/schemas/ConditionDetails'
        imei:
          type: string
          description: IMEI numarası (telefon için)
        serialNumber:
          type: string
          description: Seri numarası
        images:
          type: array
          items:
            type: string
            format: uri
          description: Kullanıcının yüklediği fotoğraflar
        inspectionImages:
          type: array
          items:
            type: string
            format: uri
          description: İnceleme sırasında çekilen fotoğraflar
        inspectionNotes:
          type: string
          description: İnceleme notları
        createdAt:
          type: string
          format: date-time
    ConditionDetails:
      type: object
      properties:
        screenCondition:
          type: string
          enum: [PERFECT, MINOR_SCRATCHES, CRACKED]
        bodyCondition:
          type: string
          enum: [PERFECT, MINOR_SCRATCHES, DENTS, MAJOR_DAMAGE]
        batteryHealth:
          type: integer
          minimum: 0
          maximum: 100
        functionalIssues:
          type: array
          items:
            type: string
            enum: [TOUCHSCREEN, BUTTONS, CAMERA, SPEAKER, MICROPHONE, WIFI, BLUETOOTH, CHARGING, FACE_ID, TOUCH_ID]
        accessories:
          type: array
          items:
            type: string
            enum: [BOX, CHARGER, CABLE, EARPHONES, MANUAL]

    # ============================================
    # VALUATION SCHEMAS
    # ============================================
    CreateValuationRequest:
      type: object
      required:
        - modelId
        - condition
      properties:
        modelId:
          type: string
          format: uuid
        variantId:
          type: string
          format: uuid
          description: Belirli varyant (storage, renk)
        condition:
          $ref: '#/components/schemas/TradeInItemCondition'
        conditionDetails:
          $ref: '#/components/schemas/ConditionDetails'
        images:
          type: array
          items:
            type: string
            format: uri
          description: Ürün fotoğrafları (opsiyonel ama değeri artırır)
        imei:
          type: string
          description: IMEI numarası (telefon için)
    Valuation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        tradeInItem:
          $ref: '#/components/schemas/TradeInItem'
        status:
          $ref: '#/components/schemas/ValuationStatus'
        estimatedValue:
          type: number
          description: Tahmini trade-in değeri (TL)
        minimumValue:
          type: number
          description: En düşük olası değer (TL)
        maximumValue:
          type: number
          description: En yüksek olası değer (TL)
        validUntil:
          type: string
          format: date-time
          description: Teklif geçerlilik süresi
        createdAt:
          type: string
          format: date-time
    ValuationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Valuation'
        pagination:
          $ref: '#/components/schemas/Pagination'
    AcceptValuationRequest:
      type: object
      required:
        - shippingMethod
        - settlementType
      properties:
        shippingMethod:
          type: string
          enum: [COURIER, DROPOFF, PICKUP]
          description: |
            - COURIER: Kargo ile gönder (ücretsiz kargo etiketi)
            - DROPOFF: Teslim noktasına bırak
            - PICKUP: Evden toplama (MVP'de seçili şehirlerde)
        pickupAddressId:
          type: string
          format: uuid
          description: PICKUP veya COURIER için adres ID
        dropoffLocationId:
          type: string
          format: uuid
          description: DROPOFF için teslim noktası ID
        settlementType:
          $ref: '#/components/schemas/SettlementType'
        applyToOrderId:
          type: string
          format: uuid
          description: ORDER_DISCOUNT seçildiyse uygulanacak sipariş ID

    # ============================================
    # TRADE-IN SCHEMAS
    # ============================================
    TradeIn:
      type: object
      properties:
        id:
          type: string
          format: uuid
        valuationId:
          type: string
          format: uuid
        valuation:
          $ref: '#/components/schemas/Valuation'
        tradeInItem:
          $ref: '#/components/schemas/TradeInItem'
        status:
          $ref: '#/components/schemas/TradeInStatus'
        shippingMethod:
          type: string
          enum: [COURIER, DROPOFF, PICKUP]
        trackingNumber:
          type: string
        shippingLabel:
          type: string
          format: uri
          description: Kargo etiketi PDF
        initialOffer:
          type: number
          description: İlk teklif (TL)
        finalOffer:
          type: number
          description: Fiziksel inceleme sonrası final teklif (TL)
        offerDifference:
          type: number
          description: İlk ve final teklif arasındaki fark
        settlementType:
          $ref: '#/components/schemas/SettlementType'
        settlementStatus:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
        appliedToOrderId:
          type: string
          format: uuid
          description: ORDER_DISCOUNT ise uygulanan sipariş
        creditTransactionId:
          type: string
          format: uuid
          description: CREDIT ise oluşturulan kredi işlemi
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEvent'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TimelineEvent:
      type: object
      properties:
        status:
          type: string
        timestamp:
          type: string
          format: date-time
        note:
          type: string
        actor:
          type: string
          enum: [SYSTEM, USER, ADMIN]
    TradeInListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TradeIn'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ============================================
    # ORDER SCHEMAS
    # ============================================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time

    # ============================================
    # ADMIN SCHEMAS (Internal Use Only)
    # ============================================
...